feat: Add support for mid-session view distance updates
Details:
- Introduced ClientInformationPlay to handle client settings updates in play state.
- Added client_information.rs with logic to:
  • Process client view distance change requests
  • Apply server limits via get_effective_render_distance()
  • Update the PlayerRenderDistance component
  • Send SetRenderDistance packet to notify the client
  • Log render distance changes for debugging
- Integrated the handler into the ECS schedule so Client Information packets
  are processed during gameplay.

This ensures that clients can adjust their view distance dynamically while the
server enforces limits and keeps both client and server render distances in sync.

+++ src/lib/net/src/packets/outgoing/set_render_distance.rs
@@ -7,7 +7,7 @@ pub struct SetRenderDistance {
     pub distance: VarInt,
 }
 
-const DEFAULT_RENDER_DISTANCE: u8 = 5;
+const DEFAULT_RENDER_DISTANCE: u32 = 5;
 
 impl Default for SetRenderDistance {
     fn default() -> Self {
@@ -16,7 +16,7 @@ impl Default for SetRenderDistance {
 }
 
 impl SetRenderDistance {
-    pub fn new(distance: u8) -> Self {
+    pub fn new(distance: u32) -> Self {
         Self {
             distance: VarInt::new(distance as i32),
         }
